Consolidate FactoryDefinition into Factory module
Fixes missing export of FactoryDefinition from package.
