--- src/factory-definition.ts
@@ -1,3 +1,0 @@
-export interface FactoryDefinition<T> {
-  create(injections?: Object): T;
-}
